useActionStack moved to its own file so it can be used anywhere
